Coale Demeny life expectancy

Life expectancy and ratio of 30d10 to 20d40 from Coale Demeny life tables, to be used for approximation of open age interval life expectancy, which is an input to SEG DDM estimation. Life tables are downloaded using the demogR package.

Usage

coale_demeny_ex

Format

[data.table()] columns:

cd_region

[character()]
"north", "south", "west", or "east"

sex

[character()]
"male", "female", or "all

index

[numeric()]
Life table level, based on life expectancy at age 10

age_start

[numeric()]
Lower bound of age interval

ex

[numeric()]
Life expectancy

ratio_30d10_20d40

[numeric()]
Ratio 30d10 to 20d40

Examples

coale_demeny_ex
#>       cd_region    sex index age_start        ex ratio_30d10_20d40
#>          <char> <char> <int>     <num>     <num>             <num>
#>    1:     north   male     1         0 17.551426         1.3585262
#>    2:     north   male     1         1 26.727418         1.3585262
#>    3:     north   male     1         5 34.027852         1.3585262
#>    4:     north   male     1        10 33.687842         1.3585262
#>    5:     north   male     1        15 30.840798         1.3585262
#>   ---                                                             
#> 6296:      west    all    25        75  9.621145         0.1354974
#> 6297:      west    all    25        80  7.053336         0.1354974
#> 6298:      west    all    25        85  4.976275         0.1354974
#> 6299:      west    all    25        90  3.385984         0.1354974
#> 6300:      west    all    25        95  2.228190         0.1354974