Estimate average annual deaths between censuses

Use annual death counts for all or some years between two censuses to approximate average annual deaths in the intercensal period. Used to prep input data for death distribution methods.

Usage

deaths_per_year(
  years_deaths,
  input_deaths,
  years_census,
  step_size = 0.001,
  rule = 2,
  ...
)

Arguments

years_deaths

[numeric()]
Which years reported deaths correspond to.

input_deaths

[numeric()]
How many deaths were recorded for each of the years in years_deaths.

years_census

[numeric(2)]
Vector of two numerics, where each number is a census year, and together they form the bounds of the intercensal period.

step_size

[numeric()]
Interpolate annual deaths at intervals of step_size. Default is 0.001 and annual deaths at intervals of 0.001 years will be approximated with interpolation.

rule

[numeric()]
Rule to be passed to stats::approxfun for extrapolation. Default is '2' and constant extrapolation from the extremes is used.

...

Other arguments to be passed to stats::approxfun.

Value

[numeric(1)]
Average annual deaths in the period.

Details

If only one year of death counts is available, that count is returned as the average annual death count. If multiple years are available, linear interpolation is used to approximate annual death counts at all points in the intercensal period, and then a simple average is taken. Constant extrapolation is used at the bounds (stats::approxfun rule = 2) by default.

Examples

annual_deaths <- deaths_per_year(
  years_deaths = c(2000, 2005, 2006),
  input_deaths = c(100, 110, 114),
  years_census = c(2000, 2010)
)