General Growth Balance (GGB) DDM

Estimate completeness of death registration using two censuses with age-specific population, and average annual age-specific deaths recorded between the censuses, following the General Growth Balance method.

Usage

ggb(
  dt,
  age_trim_lower = 5,
  age_trim_upper = 75,
  drop_open_interval = T,
  id_cols = "age_start",
  migration = F,
  input_deaths_annual = T,
  input_migrants_annual = T,
  method_growth_rate = "hyc"
)

Arguments

dt

[data.table()]
Input data, must include columns 'pop1', 'pop2', 'deaths', 'age_start', 'date1', 'date2', and all id_cols. Optional column 'migrants' to be included if migration = T.

age_trim_lower

[numeric(1)]
age-trim: lower bound of 'age_start' included (inclusive)

age_trim_upper

[numeric(1)]
age-trim: upper bound of 'age_start' included (exclusive)

drop_open_interval

[logical(1)]
Whether to drop the open age interval if it falls within the age trim bounds.

id_cols

[character()]
Column names that uniquely identify rows of dt. One column must be 'age_start'. When 'age_start' is removed, these columns define the groups to calculate completeness for.

migration

[logical(1)]
Whether a 'migrants' column should be used. If migration is available and of reasonable quality, it improves the DDM estimation. However, these data are not often available.

input_deaths_annual

[logical(1)]
Whether input deaths are formatted as annual average recorded per year. Alternative is recorded deaths for the entire period between census 1 and census 2. Default TRUE and assumes deaths are annual. If FALSE, deaths will be divided by the decimal number of years between 'date1' and 'date2'.

input_migrants_annual

[logical(1)]
Whether input migrants are formatted as annual average recorded per year. Alternative is recorded net migrants for the entire period between census 1 and census 2. Default TRUE and assumes net migrants are annual. If FALSE, migrants will be divided by the decimal number of years between 'date1' and 'date2'.

method_growth_rate

[character(1)]
Method to use for computation of growth rate. One of:

"hyc": Equation 8c from Hill, You, Choi (2009). \(r(a+) = (1 / t) * ln(N2(a+) / N1(a+))\).

"tr_iussp": Equation from IUSSP website and Tim Riffe DDM package. \(r(a+) = (1 / t) * (N2(a+) - N1(a+)) / (exp((log(N1(a+)) + log(N2(a+))) / 2))\).

Value

[list(2)]

dt [data.table()] Input dt returned with additional variables computed in GGB estimation, such as birthdays age a, population age a plus, entry rate, growth rate, and partial death rate.

completeness [data.table()] Slope, intercept, and estimated completeness by unique combination of id_cols.

Details

Columns of dt:

'pop1': age-specific population recorded at census 1

'pop2': age-specific population recorded at census 2

'deaths': average annual deaths recorded between census 1 and census 2, or, if 'input_deaths_annual' is FALSE, total deaths recorded between census 1 and census 2

'age_start': lower bound of age interval

'date1': date of first census, format yyyy-mm-dd

'date2': date of second census, format yyyy-mm-dd

'migrants' (optional): average annual net migrants between census 1 and census 2, or, if 'input_migrants_annual' is FALSE, total net migrants recorded between census 1 and census 2

References

Methods are based on the following sources:

Bennett NG, Horiuchi S. Mortality estimation from registered deaths in less developed countries. Demography. 1984 May;21(2):217-33.

Dorrington RE. 2013. "The Generalized Growth Balance Method". In Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). Tools for Demographic Estimation. Paris: International Union for the Scientific Study of Population. http://demographicestimation.iussp.org/content/generalized-growth-balance-method. Accessed May 27 2021.

Dorrington RE. 2013. "Synthetic extinct generations methods". In Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). Tools for Demographic Estimation. Paris: International Union for the Scientific Study of Population. http://demographicestimation.iussp.org/content/synthetic-extinct-generations-methods. Accessed May 27 2021.

Hill K, You D, Choi Y. Death distribution methods for estimating adult mortality: sensitivity analysis with simulated data errors. Demographic Research. 2009 Jul 1;21:235-54.

Murray CJ, Rajaratnam JK, Marcus J, Laakso T, Lopez AD. What can we conclude from death registration? Improved methods for evaluating completeness. PLoS Med. 2010 Apr 13;7(4):e1000262.

Tim Riffe, Everton Lima, and Bernardo Queiroz (2017). DDM: Death Registration Coverage Estimation. R package version 1.0.0. https://CRAN.R-project.org/package=DDM

Examples

library(data.table)
dt <- copy(zaf_2001_2007)
id_cols <- c("location", "sex", "age_start")
age_trim_lower = 5
age_trim_upper = 85
results <- ggb(
  dt,
  age_trim_lower = age_trim_lower,
  age_trim_upper = age_trim_upper,
  id_cols = id_cols,
  migration = TRUE,
  input_deaths_annual = FALSE,
  input_migrants_annual = FALSE
)