General Growth Balance - Synthetic Extinct Generations (GGBSEG) DDM

Estimate completeness of death registration using two censuses with age-specific population, and average annual age-specific deaths recorded between the censuses, following the General Growth Balance & Synthetic Extinct Generations hybrid method.

Usage

ggbseg(
  dt,
  age_trim_lower_ggb = 5,
  age_trim_upper_ggb = 75,
  age_trim_lower_seg = 45,
  age_trim_upper_seg = 90,
  drop_open_interval_ggb = T,
  drop_open_interval_seg = T,
  id_cols = c("age_start", "sex"),
  migration = F,
  input_deaths_annual = T,
  input_migrants_annual = T,
  method_growth_rate = "hyc",
  method_numerator_denominator_type = "Nx"
)

Arguments

dt

[data.table()]
Input data, must include columns 'pop1', 'pop2', 'deaths', 'age_start', 'sex', 'date1', 'date2', 'cd_region', and all id_cols. Optional column 'migrants' to be included if migration = T.

age_trim_lower_ggb

[numeric(1)]
age-trim: lower bound of 'age_start' included for GGB (inclusive)

age_trim_upper_ggb

[numeric(1)]
age-trim: upper bound of 'age_start' included for GGB (exclusive)

age_trim_lower_seg

[numeric(1)]
age-trim: lower bound of 'age_start' included for SEG (inclusive). Defaults to be the same as GGB lower age trim.

age_trim_upper_seg

[numeric(1)]
age-trim: upper bound of 'age_start' included for SEG (exclusive). defaults to be the same as GGB upper age trim.

drop_open_interval_ggb

[logical(1)]
Whether to drop the open age interval if it falls within the age trim bounds (for GGB component).

drop_open_interval_seg

[logical(1)]
Whether to drop the open age interval if it falls within the age trim bounds (for SEG component).

id_cols

[character()]
Column names that uniquely identify rows of dt. Must include 'age_start' and 'sex'. When 'age_start' is removed, these columns define the groups to calculate completeness for.

migration

[logical(1)]
Whether a 'migrants' column should be used. If migration is available and of reasonable quality, it improves the DDM estimation. However, these data are not often available.

input_deaths_annual

[logical(1)]
Whether input deaths are formatted as annual average recorded per year. Alternative is recorded deaths for the entire period between census 1 and census 2. Default TRUE and assumes deaths are annual. If FALSE, deaths will be divided by the decimal number of years between 'date1' and 'date2'.

input_migrants_annual

[logical(1)]
Whether input migrants are formatted as annual average recorded per year. Alternative is recorded net migrants for the entire period between census 1 and census 2. Default TRUE and assumes net migrants are annual. If FALSE, migrants will be divided by the decimal number of years between 'date1' and 'date2'.

method_growth_rate

[character(1)]
Method to use for computation of growth rate. One of:

"hyc": Equation 8c from Hill, You, Choi (2009). \(r(a+) = (1 / t) * ln(N2(a+) / N1(a+))\).

"tr_iussp": Equation from IUSSP website and Tim Riffe DDM package. \(r(a+) = (1 / t) * (N2(a+) - N1(a+)) / (exp((log(N1(a+)) + log(N2(a+))) / 2))\).

method_numerator_denominator_type

[character()]
Completeness point estimates are population estimated from deaths divided by population estimated from censuses. This setting dictates the ages the population belongs to. Must be one of:

"Nx" : numerator and denominator are both estimated population aged x. This is the default and matches the GBD methodology.

"nNx" : numerator and denominator are both estimated population aged x to x+n. This matches Tim Riffe DDM package and IUSSP methods.

Value

[list(2)]

dt [data.table()] Input dt returned with additional variables computed in GGBSEG estimation, such as birthdays age a, population age a plus, and synthetic cohort population age a plus.

completeness [data.table()] Estimated completeness by unique combination of id_cols.

Details

Columns of dt:

'pop1': age-specific population recorded at census 1

'pop2': age-specific population recorded at census 2

'deaths': average annual deaths recorded between census 1 and census 2

'age_start': lower bound of age interval

'sex': 'male', 'female', or 'all', used for matching Coale-Demeny life tables

'date1': date of first census, format yyyy-mm-dd

'date2': date of second census, format yyyy-mm-dd

'cd_region': 'north', 'south', 'east', or 'west', for the Coale-Demeny region to use for approximation of open age interval life expectancy. Suggest to default to 'west'.

'migrants' (optional): average annual net migrants between census 1 and census 2

References

Methods are based on the following sources:

Bennett NG, Horiuchi S. Mortality estimation from registered deaths in less developed countries. Demography. 1984 May;21(2):217-33.

Dorrington RE. 2013. "The Generalized Growth Balance Method". In Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). Tools for Demographic Estimation. Paris: International Union for the Scientific Study of Population. http://demographicestimation.iussp.org/content/generalized-growth-balance-method. Accessed May 27 2021.

Dorrington RE. 2013. "Synthetic extinct generations methods". In Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). Tools for Demographic Estimation. Paris: International Union for the Scientific Study of Population. http://demographicestimation.iussp.org/content/synthetic-extinct-generations-methods. Accessed May 27 2021.

Hill K, You D, Choi Y. Death distribution methods for estimating adult mortality: sensitivity analysis with simulated data errors. Demographic Research. 2009 Jul 1;21:235-54.

Murray CJ, Rajaratnam JK, Marcus J, Laakso T, Lopez AD. What can we conclude from death registration? Improved methods for evaluating completeness. PLoS Med. 2010 Apr 13;7(4):e1000262.

Tim Riffe, Everton Lima, and Bernardo Queiroz (2017). DDM: Death Registration Coverage Estimation. R package version 1.0.0. https://CRAN.R-project.org/package=DDM

Examples

library(data.table)
dt <- copy(zaf_2001_2007)
dt[, cd_region := "west"]
#>         location    pop1       pop2    deaths  migrants age_start    sex
#>           <char>   <num>      <num>     <num>     <num>     <num> <char>
#>  1: South Africa 2223006 2505743.50 197911.88 10605.138         0   male
#>  2: South Africa 2425066 2560641.90  15566.49  2848.486         5   male
#>  3: South Africa 2518985 2452338.80  11207.11  5152.799        10   male
#>  4: South Africa 2453156 2553292.90  25472.64 16573.504        15   male
#>  5: South Africa 2099417 2362519.00  54959.64 14802.771        20   male
#>  6: South Africa 1899275 2033165.30 102802.12  4714.013        25   male
#>  7: South Africa 1594624 1875482.60 145588.04 13330.725        30   male
#>  8: South Africa 1441657 1548185.20 145899.96  9692.869        35   male
#>  9: South Africa 1233813 1306899.80 135936.21  7463.925        40   male
#> 10: South Africa  967744 1104293.70 121009.82  8719.168        45   male
#> 11: South Africa  769627  888042.40 111156.87  9412.840        50   male
#> 12: South Africa  552402  708812.30  96854.39  4640.031        55   male
#> 13: South Africa  444592  491870.60  89930.45  5081.173        60   male
#> 14: South Africa  304835  394304.60  82842.80  4922.185        65   male
#> 15: South Africa  232604  241976.30  73036.32  4333.950        70   male
#> 16: South Africa  136466  163112.36  63870.64  2980.399        75   male
#> 17: South Africa   90856   87698.19  48162.99  1662.305        80   male
#> 18: South Africa   45920   70299.38  46195.66  2009.442        85   male
#>          date1      date2 cd_region
#>         <Date>     <Date>    <char>
#>  1: 2001-10-10 2007-02-15      west
#>  2: 2001-10-10 2007-02-15      west
#>  3: 2001-10-10 2007-02-15      west
#>  4: 2001-10-10 2007-02-15      west
#>  5: 2001-10-10 2007-02-15      west
#>  6: 2001-10-10 2007-02-15      west
#>  7: 2001-10-10 2007-02-15      west
#>  8: 2001-10-10 2007-02-15      west
#>  9: 2001-10-10 2007-02-15      west
#> 10: 2001-10-10 2007-02-15      west
#> 11: 2001-10-10 2007-02-15      west
#> 12: 2001-10-10 2007-02-15      west
#> 13: 2001-10-10 2007-02-15      west
#> 14: 2001-10-10 2007-02-15      west
#> 15: 2001-10-10 2007-02-15      west
#> 16: 2001-10-10 2007-02-15      west
#> 17: 2001-10-10 2007-02-15      west
#> 18: 2001-10-10 2007-02-15      west
id_cols <- c("location", "sex", "age_start")
age_trim_lower = 25
age_trim_upper = 65
results <- ggbseg(
  dt,
  age_trim_lower_ggb = age_trim_lower,
  age_trim_upper_ggb = age_trim_upper,
  id_cols = id_cols,
  migration = FALSE,
  input_deaths_annual = FALSE,
  input_migrants_annual = FALSE
)